requirements.txt
Pillow==10.4.0
streamlit==1.36.0
streamlit_ext==0.1.10
